+++ manim_slides/convert.py
@@ -390,7 +390,7 @@ class RevealJS(Converter):
     def open(self, file: Path) -> None:
         webbrowser.open(file.absolute().as_uri())
 
-    def convert_to(self, dest: Path) -> None:
+    def convert_to(self, dest: Path) -> None:  # noqa: C901
         """
         Convert this configuration into a RevealJS HTML presentation, saved to
         DEST.
@@ -616,7 +616,7 @@ class PowerPoint(Converter):
 
 
 def show_config_options(function: Callable[..., Any]) -> Callable[..., Any]:
-    """Wrap a function to add a `--show-config` option."""
+    """Wrap a function to add a '--show-config' option."""
 
     def callback(ctx: Context, param: Parameter, value: bool) -> None:
         if not value or ctx.resilient_parsing:
@@ -647,7 +647,7 @@ def show_config_options(function: Callable[..., Any]) -> Callable[..., Any]:
 
 
 def show_template_option(function: Callable[..., Any]) -> Callable[..., Any]:
-    """Wrap a function to add a `--show-template` option."""
+    """Wrap a function to add a '--show-template' option."""
 
     def callback(ctx: Context, param: Parameter, value: bool) -> None:
         if not value or ctx.resilient_parsing:
@@ -699,7 +699,7 @@ def show_template_option(function: Callable[..., Any]) -> Callable[..., Any]:
     multiple=True,
     callback=validate_config_option,
     help="Configuration options passed to the converter. "
-    "E.g., pass ``-cslide_number=true`` to display slide numbers.",
+    "E.g., pass '-cslide_number=true' to display slide numbers.",
 )
 @click.option(
     "--use-template",
@@ -707,7 +707,13 @@ def show_template_option(function: Callable[..., Any]) -> Callable[..., Any]:
     metavar="FILE",
     type=click.Path(exists=True, dir_okay=False, path_type=Path),
     help="Use the template given by FILE instead of default one. "
-    "To echo the default template, use ``--show-template``.",
+    "To echo the default template, use '--show-template'.",
+)
+@click.option(
+    "--offline",
+    is_flag=True,
+    help="Download any remote content and store it in the assets folder. "
+    "The is a convenient alias to '-coffline=true'.",
 )
 @show_template_option
 @show_config_options
@@ -720,6 +726,7 @@ def convert(
     open_result: bool,
     config_options: dict[str, str],
     template: Optional[Path],
+    offline: bool,
 ) -> None:
     """Convert SCENE(s) into a given format and writes the result in DEST."""
     presentation_configs = get_scenes_presentation_config(scenes, folder)
@@ -736,6 +743,13 @@ def convert(
                 cls = RevealJS
         else:
             cls = Converter.from_string(to)
+
+        if (
+            offline
+            and issubclass(cls, (RevealJS, HtmlZip))
+            and "offline" not in config_options
+        ):
+            config_options["offline"] = "true"
 
         converter = cls(
             presentation_configs=presentation_configs,