+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (unreleased)=
 ## [Unreleased](https://github.com/jeertmans/manim-slides/compare/v5.1.9...HEAD)
 
+(unreleased-added)=
+### Added
+
+- Added `--offline` option to `manim-slides convert` for offline
+  HTML presentations.
+  [#440](https://github.com/jeertmans/manim-slides/pull/440)
+
 (v5.1.9)=
 ## [v5.1.9](https://github.com/jeertmans/manim-slides/compare/v5.1.8...v5.1.9)